temporal solution on update and input types

+++ src/main/kotlin/com/github/jmlb23/movielense/repositories/RatingRepository.kt
@@ -9,7 +9,7 @@ import org.joda.time.DateTime
 object RatingRepository : Repository<Rating>{
 
     override fun filter(predicate: SqlExpressionBuilder.() -> Op<Boolean>): Sequence<Rating> = transactionEnviroment {
-        Ratings.select(predicate).map{ Rating(it[Ratings.id].toLong(),it[Ratings.userId].toLong(),it[Ratings.movieId].toLong(),it[Ratings.rating].toByte(),it[Ratings.ratedAt].toDate())}
+        Ratings.select(predicate).map{ Rating(it[Ratings.id].toLong(),it[Ratings.userId].toLong(),it[Ratings.movieId].toLong(),it[Ratings.rating],it[Ratings.ratedAt].toDate())}
 
     }.asSequence()
 
@@ -17,7 +17,7 @@ object RatingRepository : Repository<Rating>{
             transactionEnviroment { Ratings.insert {
                 it[movieId]= element.movieId.toInt()
                 it[userId] = element.userId.toInt()
-                it[rating] = element.rating.toInt()
+                it[rating] = element.rating
                 it[ratedAt] = DateTime(element.ratedAt)
             }.generatedKey!!.toLong()
         }
@@ -27,7 +27,7 @@ object RatingRepository : Repository<Rating>{
         transactionEnviroment {
             Ratings.selectAll()
                     .toList()
-                    .map{ Rating(it[Ratings.id].toLong(),it[Ratings.userId].toLong(),it[Ratings.movieId].toLong(),it[Ratings.rating].toByte(),it[Ratings.ratedAt].toDate())}
+                    .map{ Rating(it[Ratings.id].toLong(),it[Ratings.userId].toLong(),it[Ratings.movieId].toLong(),it[Ratings.rating],it[Ratings.ratedAt].toDate())}
         }.asSequence()
 
 
@@ -42,7 +42,7 @@ object RatingRepository : Repository<Rating>{
             Ratings.update({Ratings.id eq indexer.toInt()}) {
                 it[movieId]= element.movieId.toInt()
                 it[userId] = element.userId.toInt()
-                it[rating] = element.rating.toInt()
+                it[rating] = element.rating
                 it[ratedAt] = DateTime(element.ratedAt)
             }
         }.toLong()
@@ -52,7 +52,7 @@ object RatingRepository : Repository<Rating>{
         transactionEnviroment {
             Ratings
                     .select {Ratings.id eq indexer.toInt()}
-                    .map{ Rating(it[Ratings.id].toLong(),it[Ratings.userId].toLong(),it[Ratings.movieId].toLong(),it[Ratings.rating].toByte(),it[Ratings.ratedAt].toDate())}
+                    .map{ Rating(it[Ratings.id].toLong(),it[Ratings.userId].toLong(),it[Ratings.movieId].toLong(),it[Ratings.rating],it[Ratings.ratedAt].toDate())}
                     .first()
         }
